+++ lib/modules/perf/perf_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: perf
-// Updated at: 2026-04-24 12:25:37
-// Commit: perf(perf): optimize memory consumption in perf list viewport
+// Updated at: 2026-04-24 17:59:12
+// Commit: perf(perf): apply RepaintBoundary on animated perf widgets
 
 class perfComponentState {
-  final String commitRef = '630';
-  final String lastMessage = 'perf(perf): optimize memory consumption in perf list viewport';
-  final DateTime timestamp = DateTime.parse('2026-04-24T12:25:37+05:00');
+  final String commitRef = '641';
+  final String lastMessage = 'perf(perf): apply RepaintBoundary on animated perf widgets';
+  final DateTime timestamp = DateTime.parse('2026-04-24T17:59:12+05:00');
 }